Fixed some variable handling
Was originally loading user input directly into the variables that were passed to the ed25519* functions. This was not working as the strings needed to be in the appropriate type defs. I am now doing an strncpy into a dedicated variable from the user input.

Also updated the test input. It should now pass all tests and the test input is internally consistent. I need to verify these values against another library before I completely trust it.